Move keybindings out of Sway and back into GTK
Using `bindsym exec` in Sway is too slow and causes too much CPU usage
for rapid-fire keybinds.

We overcome the issue of the compositor swallowing keypresses intended
for the window switcher by stealthily changing the Sway keybind mode
when we open the switcher, and then changing it back when we're done.